Add state based priorities, for prioritizing which will draw important enough states(*) *See StatePriorities_t in info.c, it seems a bit odd to see corpses disappearing, but they are one of the first things to go when out of sprites

+++ src/info.h
@@ -1187,7 +1187,7 @@ typedef struct
 	actionf_t action;
 	statenum_t nextstate;
 	
-	int Priority;					// View priority of the state
+	uint8_t Priority;					// View priority of the state
 } state_t;
 
 extern state_t states[NUMSTATES];
